Renamed package for resilience parts #2280
