+++ sechub-scan-product-checkmarx/src/test/java/com/mercedesbenz/sechub/domain/scan/product/checkmarx/CheckmarxBadRequestConsultantTest.java
@@ -9,10 +9,10 @@ import org.junit.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.client.HttpClientErrorException;
 
+import com.mercedesbenz.sechub.commons.core.resilience.ResilienceContext;
+import com.mercedesbenz.sechub.commons.core.resilience.ResilienceProposal;
+import com.mercedesbenz.sechub.commons.core.resilience.RetryResilienceProposal;
 import com.mercedesbenz.sechub.domain.scan.SecHubExecutionException;
-import com.mercedesbenz.sechub.sharedkernel.resilience.ResilienceContext;
-import com.mercedesbenz.sechub.sharedkernel.resilience.ResilienceProposal;
-import com.mercedesbenz.sechub.sharedkernel.resilience.RetryResilienceProposal;
 
 public class CheckmarxBadRequestConsultantTest {
